test passes but fails intermittently lots of FIXMEs
