+++ hapi-fhir-jpaserver-mdm/src/main/java/ca/uhn/fhir/jpa/mdm/broker/MdmMessageHandler.java
@@ -67,8 +67,6 @@ public class MdmMessageHandler implements MessageHandler {
 	private IMdmSettings myMdmSettings;
 	@Autowired
 	private IMdmModelConverterSvc myModelConverter;
-	@Autowired
-	private MdmMessageHasher myMessageHasher;
 
 	@Override
 	public void handleMessage(Message<?> theMessage) throws MessagingException {
@@ -118,7 +116,6 @@ public class MdmMessageHandler implements MessageHandler {
 			IBaseResource targetResource = theMsg.getPayload(myFhirContext);
 			ResourceOperationMessage outgoingMsg = new ResourceOperationMessage(myFhirContext, targetResource, theMsg.getOperationType());
 			outgoingMsg.setTransactionId(theMsg.getTransactionId());
-			outgoingMsg.setMessageHash(myMessageHasher.getMessageHashOrNull(targetResource));
 
 			MdmLinkEvent linkChangeEvent = new MdmLinkEvent();
 			mdmContext.getMdmLinks()
