+++ hapi-fhir-jpaserver-subscription/src/main/java/ca/uhn/fhir/jpa/subscription/api/ISubscriptionMessageHasher.java
@@ -1,0 +1,7 @@
+package ca.uhn.fhir.jpa.subscription.api;
+
+import org.hl7.fhir.instance.model.api.IBaseResource;
+
+public interface ISubscriptionMessageHasher {
+	Integer getMessageHashOrNull(IBaseResource thePayloadResource);
+}
