+++ hapi-fhir-server/src/main/java/ca/uhn/fhir/rest/server/messaging/json/ResourceOperationJsonMessage.java
@@ -61,14 +61,6 @@ public class ResourceOperationJsonMessage extends BaseJsonMessage<ResourceOperat
 	}
 
 	@Override
-	public Integer getMessageHashOrNull() {
-		if (myPayload == null) {
-			return null;
-		}
-		return myPayload.getMessageHashOrNull();
-	}
-
-	@Override
 	public String toString() {
 		return new ToStringBuilder(this)
 			.append("myPayload", myPayload)
