+++ hapi-fhir-storage/src/main/java/ca/uhn/fhir/jpa/subscription/model/ResourceModifiedJsonMessage.java
@@ -53,6 +53,14 @@ public class ResourceModifiedJsonMessage extends BaseJsonMessage<ResourceModifie
 	}
 
 	@Override
+	public Integer getMessageHashOrNull() {
+		if (myPayload == null) {
+			return null;
+		}
+		return myPayload.getMessageHashOrNull();
+	}
+
+	@Override
 	public String toString() {
 		return new ToStringBuilder(this)
 			.append("myPayload", myPayload)
